Noted severe change.
These changes will most likely break most code that relies on this library.
Using generators is great and all, but if you want to query Wolfram Alpha for more complex information with
more context this library needs to be able to do that and be expandable at the same time. There's a lot more to Wolfram Alpha and the Python API should be able to access that content in a uniform way.