Explained limitations, offered alternative
Added a section at the bottom explaining what a developer should do if they need access to information that the Wolfram Alpha API provides but that isn't built into the interface. Because of the way this update handles the API's replies the information is easy to access using builtin data types.