It's a change. :D
Contrary to what I thought, this update appears to be mostly compatible with all the code listed that came before. All generators have been switched out for lists so that information can be used from one location. If this is determined to somehow have an impact despite Wolfram Alpha not being expected to return oodles of information, then the code can be easily rewritten to use generators instead of lists. Lists are simply more convenient when experimenting with the source code.